+++ src/main/java/br/com/postech_farmabusca/core/service/PharmacyService.java
@@ -1,10 +1,16 @@
 package br.com.postech_farmabusca.core.service;
 
+import br.com.postech_farmabusca.commoms.exception.ForbiddenException;
 import br.com.postech_farmabusca.commoms.exception.NotFoundException;
 import br.com.postech_farmabusca.commoms.mappers.PharmacyMapper;
 import br.com.postech_farmabusca.core.domain.Pharmacy;
+import br.com.postech_farmabusca.core.domain.User;
+import br.com.postech_farmabusca.core.domain.UserType;
+import br.com.postech_farmabusca.resources.entities.UserEntity;
 import br.com.postech_farmabusca.resources.repository.PharmacyRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -18,7 +24,10 @@ public class PharmacyService {
     private final PharmacyRepository pharmacyRepository;
     private final PharmacyMapper mapper;
 
+    @Transactional
     public Pharmacy createPharmacy(Pharmacy pharmacy) {
+        ensureAdminAccess();
+
         return mapper.toDomain(pharmacyRepository.save(mapper.toEntity(pharmacy)));
     }
 
@@ -36,6 +45,8 @@ public class PharmacyService {
 
     @Transactional
     public Pharmacy updatePharmacy(Long id, Pharmacy updatedPharmacy) {
+        ensureAdminAccess();
+
         Pharmacy pharmacy = getPharmacy(id);
 
         if (updatedPharmacy.getName() != null) {
@@ -50,7 +61,26 @@ public class PharmacyService {
 
     @Transactional
     public void deletePharmacy(Long id) {
+        ensureAdminAccess();
+
         Pharmacy pharmacy = getPharmacy(id);
         pharmacyRepository.deleteById(pharmacy.getId());
+    }
+
+    private void ensureAdminAccess() {
+        User user = getAuthenticatedUser();
+        if (!user.getRole().equals(UserType.ADMIN)) {
+            throw new ForbiddenException("Acesso negado: Apenas administradores podem executar esta ação.");
+        }
+    }
+
+    private User getAuthenticatedUser() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+
+        if (authentication == null || !authentication.isAuthenticated()) {
+            throw new ForbiddenException("Usuário não autenticado. Faça login como ADMIN para continuar.");
+        }
+
+        return (User) authentication.getPrincipal();
     }
 }
